Move shader into shaders module
