+++ packages/vfx/src/ParticlesMaterial.tsx
@@ -2,7 +2,7 @@ import React, { forwardRef, useMemo } from "react"
 import { AddEquation, CustomBlending } from "three"
 import CustomShaderMaterial, { iCSMProps } from "three-custom-shader-material"
 import CustomShaderMaterialImpl from "three-custom-shader-material/vanilla"
-import { createShader } from "./shader"
+import { createShader } from "./shaders/shader"
 
 type ParticlesMaterialProps = Omit<iCSMProps, "ref"> & {
   billboard?: boolean
